fix lint errors and build issues
