+++ src/lib/prisma.ts
@@ -1,5 +1,11 @@
 // src/lib/prisma.ts
 import { PrismaClient } from "@prisma/client";
-const g = global as any;
-export const prisma: PrismaClient = g.prisma || new PrismaClient();
-if (process.env.NODE_ENV !== "production") g.prisma = prisma;
+
+const globalForPrisma = globalThis as unknown as {
+  prisma?: PrismaClient;
+};
+
+export const prisma: PrismaClient =
+  globalForPrisma.prisma ?? new PrismaClient();
+
+if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
